Expand chef-utils yard comments and make consistent
I'm using these to generate the vscode snippets now so the comment is what we display to the end user. I aligned everything that was platform_family to be a platform family check. Some of them are kinda silly since they almost entirely 1:1 match, but this way it's all consistent. I'm not too concerned about someone with omnios in vscode at this point, but it's all done.

Signed-off-by: Tim Smith <tsmith@chef.io>